refactor: 생성, 수정 모달 컴포넌트 모달화 및 데이터 구조 변경에 해당하는 입력창 생성
